refactor(lib): extract shared helpers into tools/common.rs (#1192)
