+++ crates/aptu-coder/src/tools/common.rs
@@ -1,0 +1,140 @@
+// SPDX-FileCopyrightText: 2026 aptu-coder contributors
+// SPDX-License-Identifier: Apache-2.0
+//! Shared helpers used by two or more tool handler modules.
+//!
+//! Items here are `pub(crate)` and re-exported from `lib.rs` at the crate root
+//! so that integration tests can access them via `aptu_coder::` paths.
+
+use rmcp::model::{CallToolResult, Content, ErrorData, Meta};
+
+/// Returns `true` when `summary=true` and a `cursor` are both provided, which is an invalid
+/// combination since summary mode and pagination are mutually exclusive.
+#[must_use]
+pub fn summary_cursor_conflict(summary: Option<bool>, cursor: Option<&str>) -> bool {
+    summary == Some(true) && cursor.is_some()
+}
+
+/// Session and client metadata recorded as span attributes on every tool call.
+pub struct ClientMetadata {
+    pub session_id: Option<String>,
+    pub client_name: Option<String>,
+    pub client_version: Option<String>,
+}
+
+/// Extract W3C Trace Context from MCP request _meta field and set as parent span context.
+///
+/// Attempts to extract traceparent and tracestate from the request's _meta field.
+/// If successful, calls `set_parent` on the current tracing span so the OTel layer
+/// re-parents it to the caller's trace. This must be called after the `#[instrument]`
+/// span has been entered (i.e., inside the function body) for `set_parent` to take effect.
+/// If extraction fails or _meta is absent, silently proceeds with root context (no panic).
+pub fn extract_and_set_trace_context(
+    meta: Option<&rmcp::model::Meta>,
+    client_meta: ClientMetadata,
+) {
+    use tracing_opentelemetry::OpenTelemetrySpanExt as _;
+
+    let span = tracing::Span::current();
+
+    // Record session and client attributes
+    if let Some(sid) = client_meta.session_id {
+        span.record("mcp.session.id", &sid);
+    }
+    if let Some(cn) = client_meta.client_name {
+        span.record("client.name", &cn);
+    }
+    if let Some(cv) = client_meta.client_version {
+        span.record("client.version", &cv);
+    }
+
+    // Extract agent-session-id from _meta if present (opportunistic; silent no-op if absent)
+    if let Some(asi_str) = meta.and_then(|m| m.0.get("agent-session-id").and_then(|v| v.as_str())) {
+        span.record("mcp.client.session.id", asi_str);
+    }
+
+    let Some(meta) = meta else { return };
+
+    let mut propagation_map = std::collections::HashMap::new();
+
+    // Extract traceparent if present
+    if let Some(traceparent) = meta.0.get("traceparent")
+        && let Some(tp_str) = traceparent.as_str()
+    {
+        propagation_map.insert("traceparent".to_string(), tp_str.to_string());
+    }
+
+    // Extract tracestate if present
+    if let Some(tracestate) = meta.0.get("tracestate")
+        && let Some(ts_str) = tracestate.as_str()
+    {
+        propagation_map.insert("tracestate".to_string(), ts_str.to_string());
+    }
+
+    // Only attempt extraction if we have at least traceparent
+    if propagation_map.is_empty() {
+        return;
+    }
+
+    // Extract context via the globally registered propagator (TraceContextPropagator by default)
+    let parent_cx = opentelemetry::global::get_text_map_propagator(|propagator| {
+        propagator.extract(&ExtractMap(&propagation_map))
+    });
+
+    // Re-parent the current tracing span (already entered via #[instrument]) to the
+    // extracted OTel context. set_parent is a no-op if the OTel layer is not installed.
+    let _ = span.set_parent(parent_cx);
+}
+
+/// Helper struct for W3C Trace Context extraction from HashMap
+struct ExtractMap<'a>(&'a std::collections::HashMap<String, String>);
+
+impl<'a> opentelemetry::propagation::Extractor for ExtractMap<'a> {
+    fn get(&self, key: &str) -> Option<&str> {
+        self.0.get(key).map(|s| s.as_str())
+    }
+
+    fn keys(&self) -> Vec<&str> {
+        self.0.keys().map(|k| k.as_str()).collect()
+    }
+}
+
+#[derive(Debug, Clone, Copy, serde::Serialize)]
+#[serde(rename_all = "camelCase")]
+struct ErrorMeta {
+    error_category: &'static str,
+    is_retryable: bool,
+    suggested_action: &'static str,
+}
+
+#[must_use]
+pub fn error_meta(
+    category: &'static str,
+    is_retryable: bool,
+    suggested_action: &'static str,
+) -> serde_json::Value {
+    serde_json::to_value(ErrorMeta {
+        error_category: category,
+        is_retryable,
+        suggested_action,
+    })
+    .unwrap_or_default()
+}
+
+#[must_use]
+pub fn err_to_tool_result(e: ErrorData) -> CallToolResult {
+    let mut result =
+        CallToolResult::error(vec![Content::text(e.message)]).with_meta(Some(no_cache_meta()));
+    if let Some(data) = e.data {
+        result.structured_content = Some(data);
+    }
+    result
+}
+
+pub fn no_cache_meta() -> Meta {
+    let mut m = serde_json::Map::new();
+    m.insert(
+        "cache_hint".to_string(),
+        serde_json::Value::String("no-cache".to_string()),
+    );
+    Meta(m)
+}
